pick a product to be edited, put the data in edit product form,

+++ src/components/SingleProduct.js
@@ -1,18 +1,43 @@
 import { useDispatch } from "react-redux";
-import { Link } from "react-router-dom";
-import { deleteProductAction } from "../actions/productAction";
+import { Link, useHistory } from "react-router-dom";
+import {
+  deleteProductAction,
+  getEditProductAction,
+} from "../actions/productAction";
+import Swal from "sweetalert2";
 
 const SingleProduct = ({ product }) => {
   const { id, product_name, price } = product;
 
   const dispatch = useDispatch();
+  //enable history to redirection
+  const history = useHistory();
 
   //product delete confirmation
   const productDeleteOK = (id) => {
     //ask for user confirmation
+    Swal.fire({
+      title: "Are you sure?",
+      text: "You won't be able to revert this!",
+      icon: "warning",
+      showCancelButton: true,
+      confirmButtonColor: "#3085d6",
+      cancelButtonColor: "#d33",
+      confirmButtonText: "Yes, delete it!",
+      cancelButtonText: "Cancel",
+    }).then((result) => {
+      if (result.isConfirmed) {
+        //pass the id to the action to delte from DB and update the state
+        dispatch(deleteProductAction(id));
+      }
+    });
+  };
 
-    //pass the id to the action to delte from DB and update the state
-    dispatch(deleteProductAction(id));
+  //function to redirect using history
+  const redirectEdition = (product) => {
+    //dispatch to get object data to fill de edit form an then edit it
+    dispatch(getEditProductAction(product));
+    history.push(`/products/edit/${product.id}`);
   };
 
   return (
@@ -22,9 +47,13 @@ const SingleProduct = ({ product }) => {
         <span className="font-weight-bold">$ {price}</span>
       </td>
       <td className="actions">
-        <Link to={`/products/edit/${id}`} className="btn btn-primary mr-2">
+        <button
+          type="button"
+          onClick={() => redirectEdition(product)}
+          className="btn btn-primary mr-2"
+        >
           Edit
-        </Link>
+        </button>
         <button
           type="button"
           className="btn btn-danger"
